refactor(oracle): extract scapy encode helpers module
